requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
anyio==4.2.0
    # via httpx
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via typer
colorama==0.4.6
    # via typer
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via letsbuilda-pypi
idna==3.6
    # via
    #   anyio
    #   httpx
letsbuilda-pypi==5.0.2
    # via -r requirements/requirements.in
linkify-it-py==2.0.2
    # via markdown-it-py
markdown-it-py[linkify,plugins]==3.0.0
    # via
    #   markdown-it-py
    #   mdit-py-plugins
    #   rich
    #   textual
mdit-py-plugins==0.4.0
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
pygments==2.17.2
    # via rich
rich==13.7.0
    # via
    #   textual
    #   typer
shellingham==1.5.4
    # via typer
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
textual==0.48.2
    # via -r requirements/requirements.in
typer[all]==0.9.0
    # via -r requirements/requirements.in
typing-extensions==4.9.0
    # via
    #   textual
    #   typer
uc-micro-py==1.0.2
    # via linkify-it-py
xmltodict==0.13.0
    # via letsbuilda-pypi
